+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/limitreduction/DefaultLimitsReducer.java
@@ -40,7 +40,7 @@ public class DefaultLimitsReducer extends AbstractLimitsReducer<LoadingLimits> {
                 .sorted(Comparator.comparing(LoadingLimits.TemporaryLimit::getAcceptableDuration)).toList();
         double previousRetainedReducedValue = Double.NaN;
         for (LoadingLimits.TemporaryLimit tl : temporaryLimits) { // iterate in ascending order of the durations
-            float reduction = getTemporaryLimitReduction(tl.getAcceptableDuration());
+            double reduction = getTemporaryLimitReduction(tl.getAcceptableDuration());
             double tlReducedValue = applyReduction(tl.getValue(), reduction);
             if (Double.isNaN(previousRetainedReducedValue) || tlReducedValue < previousRetainedReducedValue) {
                 previousRetainedReducedValue = tlReducedValue;
@@ -58,7 +58,7 @@ public class DefaultLimitsReducer extends AbstractLimitsReducer<LoadingLimits> {
 
     private AbstractReducedLoadingLimits initReducedLoadingLimits(LimitType type, double permanentLimit,
                                                                   double originalPermanentLimit,
-                                                                  float reductionAppliedOnPermanentLimit) {
+                                                                  double reductionAppliedOnPermanentLimit) {
         return switch (type) {
             case ACTIVE_POWER -> new ReducedActivePowerLimits(permanentLimit, originalPermanentLimit, reductionAppliedOnPermanentLimit);
             case APPARENT_POWER -> new ReducedApparentPowerLimits(permanentLimit, originalPermanentLimit, reductionAppliedOnPermanentLimit);
